+++ test/block/IIf.test.js
@@ -1,4 +1,6 @@
 var assert = require('assert');
+var Promise = require('es6-promise').Promise;
+var Conversation = require('../../lib/Conversation');
 var Block = require('../../lib/block/Block');
 var IIf = require('../../lib/block/IIf');
 var Then = require('../../lib/block/Then');
@@ -6,10 +8,10 @@ var Then = require('../../lib/block/Then');
 describe('IIf', function() {
 
   it('should create an iif without trueBlock and falseBlock', function () {
-    var c = {};
+    var conversation = new Conversation();
     var condition = function (message, context) {
       assert.equal(message, 'message');
-      assert.strictEqual(context, c);
+      assert.strictEqual(context, conversation.context);
       return true;
     };
     var iif = new IIf(condition);
@@ -18,15 +20,16 @@ describe('IIf', function() {
     assert.ok(iif instanceof IIf);
 
 
-    var next = iif.execute('message', c);
-    assert.equal(next.result, 'message');
-    assert.ok(next.block instanceof Then);
+    return iif.execute(conversation, 'message').then(function (next) {
+      assert.equal(next.result, 'message');
+      assert.ok(next.block instanceof Then);
+    });
   });
 
   it('should create an iif with trueBlock, falseBlock, and next block', function () {
-    var c = {};
+    var conversation = new Conversation();
     var condition = function (message, context) {
-      assert.strictEqual(context, c);
+      assert.strictEqual(context, conversation.context);
       return message === 'yes';
     };
     var trueBlock = new Then(function () {});
@@ -36,17 +39,22 @@ describe('IIf', function() {
 
     assert.ok(iif instanceof IIf);
 
-    var next = iif.execute('yes', c);
-    assert.equal(next.result, 'yes');
-    assert.strictEqual(next.block, trueBlock);
+    return iif.execute(conversation, 'yes')
+        .then(function (next) {
+          assert.equal(next.result, 'yes');
+          assert.strictEqual(next.block, trueBlock);
 
-    next = iif.execute('no', c);
-    assert.equal(next.result, 'no');
-    assert.strictEqual(next.block, falseBlock);
+          return iif.execute(conversation, 'no');
+        })
+        .then(function (next) {
+          assert.equal(next.result, 'no');
+          assert.strictEqual(next.block, falseBlock);
+        });
+
   });
 
   it('should create an iif with a RegExp condition', function () {
-    var c = {};
+    var conversation = new Conversation();
     var condition = /yes/;
     var trueBlock = new Then(function () {});
     var falseBlock = new Then(function () {});
@@ -54,17 +62,21 @@ describe('IIf', function() {
 
     assert.ok(iif instanceof IIf);
 
-    var next = iif.execute('yes', c);
-    assert.equal(next.result, 'yes');
-    assert.strictEqual(next.block, trueBlock);
+    return iif.execute(conversation, 'yes')
+        .then(function (next) {
+          assert.equal(next.result, 'yes');
+          assert.strictEqual(next.block, trueBlock);
 
-    next = iif.execute('no', c);
-    assert.equal(next.result, 'no');
-    assert.strictEqual(next.block, falseBlock);
+          return iif.execute(conversation, 'no');
+        })
+        .then(function (next) {
+          assert.equal(next.result, 'no');
+          assert.strictEqual(next.block, falseBlock);
+        });
   });
 
   it('should create an iif with a string condition', function () {
-    var c = {};
+    var conversation = new Conversation();
     var condition = 'yes';
     var trueBlock = new Then(function () {});
     var falseBlock = new Then(function () {});
@@ -72,17 +84,49 @@ describe('IIf', function() {
 
     assert.ok(iif instanceof IIf);
 
-    var next = iif.execute('yes', c);
-    assert.equal(next.result, 'yes');
-    assert.strictEqual(next.block, trueBlock);
+    return iif.execute(conversation, 'yes')
+        .then(function (next) {
+          assert.equal(next.result, 'yes');
+          assert.strictEqual(next.block, trueBlock);
 
-    next = iif.execute('no', c);
-    assert.equal(next.result, 'no');
-    assert.strictEqual(next.block, falseBlock);
+          return iif.execute(conversation, 'no');
+        })
+        .then(function (next) {
+          assert.equal(next.result, 'no');
+          assert.strictEqual(next.block, falseBlock);
+        });
+  });
+
+  it('should create an iif with a condition returning a Promise', function () {
+    var conversation = new Conversation();
+    var condition = function (message) {
+      return new Promise(function (resolve, reject) {
+        setTimeout(function () {
+          resolve(message == 'yes');
+        }, 10)
+      })
+    };
+    var trueBlock = new Then(function () {});
+    var falseBlock = new Then(function () {});
+    var iif = new IIf(condition, trueBlock, falseBlock);
+
+    assert.ok(iif instanceof IIf);
+
+    return iif.execute(conversation, 'yes')
+        .then(function (next) {
+          assert.equal(next.result, 'yes');
+          assert.strictEqual(next.block, trueBlock);
+
+          return iif.execute(conversation, 'no');
+        })
+        .then(function (next) {
+          assert.equal(next.result, 'no');
+          assert.strictEqual(next.block, falseBlock);
+        });
   });
 
   it('should create an iif with a number condition', function () {
-    var c = {};
+    var conversation = new Conversation();
     var condition = 42;
     var trueBlock = new Then(function () {});
     var falseBlock = new Then(function () {});
@@ -90,13 +134,17 @@ describe('IIf', function() {
 
     assert.ok(iif instanceof IIf);
 
-    var next = iif.execute(42, c);
-    assert.equal(next.result, 42);
-    assert.strictEqual(next.block, trueBlock);
+    return iif.execute(conversation, 42)
+        .then(function (next) {
+          assert.equal(next.result, 42);
+          assert.strictEqual(next.block, trueBlock);
 
-    next = iif.execute(12, c);
-    assert.equal(next.result, 12);
-    assert.strictEqual(next.block, falseBlock);
+          return iif.execute(conversation, 12);
+        })
+        .then(function (next) {
+          assert.equal(next.result, 12);
+          assert.strictEqual(next.block, falseBlock);
+        });
   });
 
   it('should throw an error on invalid input arguments', function () {