+++ test/block/Then.test.js
@@ -1,4 +1,5 @@
 var assert = require('assert');
+var Conversation = require('../../lib/Conversation');
 var Then = require('../../lib/block/Then');
 
 describe('Then', function() {
@@ -14,45 +15,53 @@ describe('Then', function() {
     assert.throws(function () { new Then('bla')}, TypeError);
   });
 
-  it('should execute a Then block without arguments', function () {
+  it('should execute a Then block without message', function () {
     var action = new Then(function (response, context) {
       assert.strictEqual(response, undefined);
-      assert.strictEqual(context, undefined);
+      assert.strictEqual(context, conversation.context);
     });
 
-    var next = action.execute();
-    assert.deepEqual(next, {
-      result: undefined,
-      block: undefined
-    })
+    var conversation = new Conversation();
+    return action.execute(conversation).then(function(next) {
+      assert.deepEqual(next, {
+        result: undefined,
+        block: undefined
+      })
+    });
   });
 
   it('should execute a Then block with context', function () {
-    var context = {a: 2};
+    var conversation = new Conversation({
+      context: {a: 2}
+    });
     var action = new Then(function (response, context) {
       assert.strictEqual(response, undefined);
       assert.deepEqual(context, {a: 2});
     });
 
-    var next = action.execute(undefined, context);
-    assert.deepEqual(next, {
-      result: undefined,
-      block: undefined
-    })
+    return action.execute(conversation, undefined).then(function(next) {
+      assert.deepEqual(next, {
+        result: undefined,
+        block: undefined
+      })
+    });
   });
 
   it('should execute a Then block with context and argument', function () {
-    var context = {a: 2};
+    var conversation = new Conversation({
+      context: {a: 2}
+    });
     var action = new Then(function (response, context) {
       assert.strictEqual(response, 'hello world');
       assert.deepEqual(context, {a: 2});
     });
 
-    var next = action.execute('hello world', context);
-    assert.deepEqual(next, {
-      result: undefined,
-      block: undefined
-    })
+    return action.execute(conversation, 'hello world').then(function(next) {
+      assert.deepEqual(next, {
+        result: undefined,
+        block: undefined
+      })
+    });
   });
 
   it('should execute a Then block with next block', function () {
@@ -60,9 +69,11 @@ describe('Then', function() {
     var nextThen = new Then (function () {});
     action.then(nextThen);
 
-    var next = action.execute();
-    assert.strictEqual(next.result, undefined);
-    assert.strictEqual(next.block, nextThen);
+    var conversation = new Conversation();
+    return action.execute(conversation).then(function(next) {
+      assert.strictEqual(next.result, undefined);
+      assert.strictEqual(next.block, nextThen);
+    });
   });
 
   it('should pass the result from and to callback when executing', function () {
@@ -71,9 +82,11 @@ describe('Then', function() {
       return 'out';
     });
 
-    var next = action.execute('in');
-    assert.strictEqual(next.result, 'out');
-    assert.strictEqual(next.block, undefined);
+    var conversation = new Conversation();
+    return action.execute(conversation, 'in').then(function(next) {
+      assert.strictEqual(next.result, 'out');
+      assert.strictEqual(next.block, undefined);
+    });
   });
 
 });